src/section7/QuestionTime/requirements.txt
asgiref==3.3.1
autopep8==1.5.5
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606237165371/work
chardet @ file:///C:/ci/chardet_1607706937985/work
confusable-homoglyphs==3.2.0
cryptography @ file:///C:/ci/cryptography_1607618721363/work
defusedxml==0.7.1
Django==3.1.6
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-registration==3.1.2
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.12.2
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
oauthlib==3.1.0
Pillow==8.1.0
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==2.0.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python3-openid==3.2.0
pytz==2021.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
six @ file:///C:/ci/six_1605205404068/work
sqlparse==0.4.1
toml==0.10.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
